+++ src/ScottPlot4/ScottPlot.Tests/PlotTypes/Scatter.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using ScottPlot;
+using System.Drawing;
 
 namespace ScottPlotTests.PlotTypes
 {
@@ -66,6 +67,58 @@ namespace ScottPlotTests.PlotTypes
 
             plt.Legend();
             plt.SetAxisLimits(0, 30, 1.42, 1.62);
+            TestTools.SaveFig(plt);
+        }
+
+        [Test]
+        public void Test_Scatter_IgnoreNaN()
+        {
+            double[] xs = ScottPlot.DataGen.Consecutive(51);
+            double[] ys = ScottPlot.DataGen.Sin(51);
+
+            Random rand = new(0);
+            double[] ys2 = ScottPlot.DataGen.InsertNanRanges(ys, rand, 5);
+            Console.WriteLine(string.Join(", ", ys2.Select(x => x.ToString())));
+
+            ScottPlot.Plot plt = new(600, 400);
+            var sp1 = plt.AddScatter(xs, ys, color: Color.FromArgb(50, Color.Black), label: "original data");
+            var sp2 = plt.AddScatter(xs, ys2, color: Color.Black, label: "data with gaps");
+            plt.Legend(location: Alignment.LowerLeft);
+
+            // default behavior throws with a NaN error
+            Assert.Throws<InvalidOperationException>(() => { plt.Render(); });
+
+            // ignoring NaN points prevents the error
+            sp2.OnNaN = ScottPlot.Plottable.ResponseToNaN.Ignore;
+            plt.Title($"OnNaN = {sp2.OnNaN}");
+            Assert.DoesNotThrow(() => plt.Render());
+
+            TestTools.SaveFig(plt);
+        }
+
+        [Test]
+        public void Test_Scatter_GapNaN()
+        {
+            double[] xs = ScottPlot.DataGen.Consecutive(51);
+            double[] ys = ScottPlot.DataGen.Sin(51);
+
+            Random rand = new(0);
+            double[] ys2 = ScottPlot.DataGen.InsertNanRanges(ys, rand, 5);
+            Console.WriteLine(string.Join(", ", ys2.Select(x => x.ToString())));
+
+            ScottPlot.Plot plt = new(600, 400);
+            var sp1 = plt.AddScatter(xs, ys, color: Color.FromArgb(50, Color.Black), label: "original data");
+            var sp2 = plt.AddScatter(xs, ys2, color: Color.Black, label: "data with gaps");
+            plt.Legend(location: Alignment.LowerLeft);
+
+            // default behavior throws with a NaN error
+            Assert.Throws<InvalidOperationException>(() => { plt.Render(); });
+
+            // gapping NaN points prevents the error
+            sp2.OnNaN = ScottPlot.Plottable.ResponseToNaN.Gap;
+            plt.Title($"OnNaN = {sp2.OnNaN}");
+            Assert.DoesNotThrow(() => plt.Render());
+
             TestTools.SaveFig(plt);
         }
     }