+++ src/ScottPlot5/Sandbox/ScottPlot.Sandbox.Eto/MainWindow.cs
@@ -15,10 +15,10 @@ namespace ScottPlot.Sandbox.Eto
 
             const int N = 51;
 
-            etoPlot.Plot.Add(DebugPoint);
-            etoPlot.Plot.AddScatter(Generate.Consecutive(N), Generate.Sin(N), Colors.Blue);
-            etoPlot.Plot.AddScatter(Generate.Consecutive(N), Generate.Cos(N), Colors.Red);
-            etoPlot.Plot.AddScatter(Generate.Consecutive(N), Generate.Sin(N, 0.5), Colors.Green);
+            etoPlot.Plot.Plottables.Add(DebugPoint);
+            etoPlot.Plot.Add.Scatter(Generate.Consecutive(N), Generate.Sin(N), Colors.Blue);
+            etoPlot.Plot.Add.Scatter(Generate.Consecutive(N), Generate.Cos(N), Colors.Red);
+            etoPlot.Plot.Add.Scatter(Generate.Consecutive(N), Generate.Sin(N, 0.5), Colors.Green);
             etoPlot.Refresh();
 
             etoPlot.MouseMove += EtoPlot_MouseMove;